Crate perpetuity

Provides a trait similar to the standard Iterator, but for sequences that never terminate. In particular, Perpetuity::next_item returns Self::Item, not Option<Self::Item>. The interface to functions like successors can also be much simpler, as neither the initial value nor the callback need deal with options. For example, compare the following two expressions:

use std::iter;
use perpetuity as infi;

let powers_of_two = iter::successors(Some(1), |a| Some(a * 2));
let powers_of_two = infi::successors(1, |a| a * 2);

One common form of infinite range is sequences of IDs. For example:

let mut ids = 1..;
let thing1 = Thing { id: ids.next().unwrap() };
let thing2 = Thing { id: ids.next().unwrap() };
// ...

Using Perpetuity, we no longer need to unwrap anything:

let mut ids = infi::from_range(1..);
let thing1 = Thing { id: ids.next_item() };
let thing2 = Thing { id: ids.next_item() };

Macros

range_from 🔒

Implements Perpetuity for infinite ranges.

Structs

IntoIter

Successors

Take

Traits

Perpetuity

Like Iterator, but returns items directly, rather than Options.

Functions

assimilate

Returns the specified Perpetuity, erasing any further type information.