perpetuity

Macro range_from

macro_rules! range_from {
    ($t:ty) => { ... };
}

Implements Perpetuity for infinite ranges.

Implementation

We can’t implement traits for RangeFrom<A> as [Iterator<Item=A>] without declaring A: Step, but Step is unstable (even though it’s used in the implementation of RangeFrom, which is stable). So, we can’t impl traits for RangeFrom as Iterator in stable Rust. See also: https://stackoverflow.com/a/56986698/3116635

If we could use Step in stable Rust, the impl might look like this:

impl<A: Step> Perpetuity for RangeFrom<A> {
    type Item = A;
    fn next_item(&mut self) -> Self::Item {
        self.next().unwrap()
    }
}

So instead, we crank out implementations for specific types of RangeFrom.

TODO

Define a perpetuity::RangeFrom tpye that doesn’t require unwrap in next_item.